backend/auth-service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
motor==3.3.1
pymongo==4.5.0
pydantic==1.10.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.0.0 